Dump more info into Package.md
