+++ Sources/SourceDocsLib/DocumentationGenerator/Elements/MarkdownVariable.swift
